+++ imports/ui/transactions/TransactionRow.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';